First Iteration
Keyboard operation
Press the 'R' key:
Function: Create a new non-overlapping circle on the screen.
(But it seems not implemented)

Press the '1' key:
Function: Set the color of the ripples to red.

Press the '2' key:
Function: Set the color of the ripples to green.

Press the '3' key:
Function: Set the color of the ripples to blue.

Mouse operation
Click the mouse:
Function: Create multiple ripple effects at the center of the circle at the click position (if the click position is inside a circle).

Drag the mouse:
Function: During the dragging process, create ripples of random colors along the path of the mouse.