Second Iteration
Try changing to make an interactive animation that generates ripples when the user presses a specific key on the keyboard, and these ripples affect the color of the circle.
Effect description
Pressing the space bar will generate a new ripple at a random position, which will interact with the circle and change the color of the circle.
Clicking the mouse will generate a new ripple at the click position.
Drag the mouse to generate continuous ripples.
Clicking the center of each shape will generate ripples, and the color of the shape will change with the color of the ripple.