Last iteration update
Interactive operation (user input)

Mouse drag:

When the user holds down the mouse and drags on the canvas, new non-overlapping circles are generated at the mouse position.

Mouse double-click:

When the user double-clicks the canvas, a new ripple center is created at the double-click position and the ripple effect begins to be generated.

Reset ripple button:

There is a button in the upper left corner of the canvas. When you click the "Reset Ripple" button, the ripple effect will be reset, including the center position, ripple diameter, and ripple expansion speed.

Increase ripple speed button:

There is also a button in the upper left corner of the canvas. When you click the "Increase Ripple Speed" button, the ripple expansion speed will be increased.

Keyboard input:

Press the space bar: reset all circles on the canvas and generate new non-overlapping circles.

Press the R key: increase the ripple expansion speed.